Move config bindings into a bundle.
Issue #750
