+++ build.gradle
@@ -46,6 +46,7 @@ dependencies {
     bundle group: 'net.engio', name: 'mbassador', version: '1.3.0'
     bundle group: 'com.google.code.gson', name: 'gson', 'version': '2.5'
 
+    bundle group: 'com.dmdirc', name: 'com.dmdirc.config.binding', version: '+', changing: true
     bundle group: 'com.dmdirc', name: 'com.dmdirc.events.eventbus', version: '+', changing: true
     bundle group: 'com.dmdirc', name: 'com.dmdirc.util.io.yaml', version: '+', changing: true
