+++ src/main/java/com/dmdirc/commandparser/CommandManager.java
@@ -20,7 +20,7 @@ package com.dmdirc.commandparser;
 import com.dmdirc.GlobalWindow;
 import com.dmdirc.commandparser.commands.Command;
 import com.dmdirc.commandparser.parsers.CommandParser;
-import com.dmdirc.config.ConfigBinding;
+import com.dmdirc.config.binding.ConfigBinding;
 import com.dmdirc.interfaces.CommandController;
 import com.dmdirc.interfaces.Connection;
 import com.dmdirc.interfaces.ConnectionManager;
