+++ src/main/java/com/dmdirc/config/ConfigBinderImpl.java
@@ -17,6 +17,8 @@
 
 package com.dmdirc.config;
 
+import com.dmdirc.config.binding.ConfigBinder;
+import com.dmdirc.config.binding.ConfigBinding;
 import com.dmdirc.interfaces.config.AggregateConfigProvider;
 import com.dmdirc.interfaces.config.ConfigChangeListener;
