+++ src/main/java/com/dmdirc/logger/DiskLoggingErrorManager.java
@@ -19,8 +19,8 @@ package com.dmdirc.logger;
 
 import com.dmdirc.commandline.CommandLineOptionsModule.Directory;
 import com.dmdirc.commandline.CommandLineOptionsModule.DirectoryType;
-import com.dmdirc.config.ConfigBinder;
-import com.dmdirc.config.ConfigBinding;
+import com.dmdirc.config.binding.ConfigBinder;
+import com.dmdirc.config.binding.ConfigBinding;
 import com.dmdirc.events.ErrorEvent;
 import com.dmdirc.events.ProgramErrorEvent;
 import com.dmdirc.events.eventbus.EventBus;
