+++ src/test/java/com/dmdirc/config/ConfigBinderImplTest.java
@@ -22,6 +22,8 @@
 
 package com.dmdirc.config;
 
+import com.dmdirc.config.binding.ConfigBinder;
+import com.dmdirc.config.binding.ConfigBinding;
 import com.dmdirc.interfaces.config.AggregateConfigProvider;
 
 import org.junit.Before;
