+++ src/test/java/com/dmdirc/logger/SentryLoggingErrorManagerTest.java
@@ -22,7 +22,7 @@
 
 package com.dmdirc.logger;
 
-import com.dmdirc.config.ConfigBinder;
+import com.dmdirc.config.binding.ConfigBinder;
 import com.dmdirc.events.ProgramErrorAddedEvent;
 import com.dmdirc.events.eventbus.EventBus;
 import com.dmdirc.interfaces.config.AggregateConfigProvider;
